move csv read to first activity screen
