+++ app/src/main/java/com/reotyranny/semeru/Controller/LocationListActivity.java
@@ -39,54 +39,11 @@ public class LocationListActivity extends AppCompatActivity {
 
 
         Model mo = Model.getInstance();
-        if (mo.places.isEmpty())
-            readSDFile();
         location = mo.places;
 
         // specify an adapter (see also next example)
         mAdapter = new LocationAdapter(location);
         mRecyclerView.setAdapter(mAdapter);
-    }
-
-    public static final int KEY = 0;
-    public static final int NAME = 1;
-    public static final int LATITUDE = 2;
-    public static final int LONGITUDE = 3;
-    public static final int STREET_ADDRESS = 4;
-    public static final int CITY = 5;
-    public static final int STATE = 6;
-    public static final int ZIP = 7;
-    public static final int TYPE = 8;
-    public static final int PHONE = 9;
-    public static final int WEBSITE = 10;
-
-    private void readSDFile() {
-
-        Model model = Model.getInstance();
-
-        try {
-            //Open a stream on the raw file
-            InputStream is = getResources().openRawResource(R.raw.locations);
-            //From here we probably should call a model method and pass the InputStream
-            //Wrap it in a BufferedReader so that we get the readLine() method
-            BufferedReader br = new BufferedReader(new InputStreamReader(is, StandardCharsets.UTF_8));
-
-            String line;
-            br.readLine(); //get rid of header line
-            while ((line = br.readLine()) != null) {
-                Log.d("test-location", line);
-                String[] tokens = line.split(",");
-                int key = Integer.parseInt(tokens[KEY]);
-                float latitude = Float.parseFloat(tokens[LATITUDE]);
-                float longitude = Float.parseFloat(tokens[LONGITUDE]);
-                model.addLocation(new Location(key, tokens[NAME], longitude, latitude, tokens[STREET_ADDRESS],
-                        tokens[CITY], tokens[STATE], tokens[ZIP], tokens[TYPE], tokens[PHONE], tokens[WEBSITE]));
-            }
-            br.close();
-        } catch (IOException e) {
-            Log.e("test-location", "error reading assets", e);
-        }
-
     }
 
 }
